Move: 단일 entity posts폴더로 이동
