+++ src/chats/dto/createChatComplaints.dto.ts
@@ -1,7 +1,7 @@
 import { Field, InputType } from '@nestjs/graphql';
 import { IsNotEmpty, IsNumber, IsString } from 'class-validator';
 import { Chat } from '../chat.entity';
-import { ComplaintReason } from 'src/complaintReasons/complaintReason.entity';
+import { ComplaintReason } from 'src/posts/entities/complaintReason.entity';
 
 @InputType()
 export class CreateChatComplaintsDto {
