+++ src/chats/dto/createChatRoom.dto.ts
@@ -1,6 +1,6 @@
 import { Field, InputType } from '@nestjs/graphql';
 import { IsNotEmpty, IsNumber } from 'class-validator';
-import { Post } from 'src/posts/post.entity';
+import { Post } from 'src/posts/entities/post.entity';
 
 @InputType()
 export class CreateChatRoomDto {
