+++ src/configs/typeorm.config.ts
@@ -1,14 +1,14 @@
 import { TypeOrmModuleOptions } from '@nestjs/typeorm';
 import * as config from 'config';
-import { Category } from 'src/categories/category.entity';
-import { DealState } from 'src/dealStates/dealState.entity';
-import { Post } from 'src/posts/post.entity';
-import { TownRange } from 'src/townRanges/townRange.entity';
+import { Category } from 'src/posts/entities/category.entity';
+import { DealState } from 'src/posts/entities/dealState.entity';
+import { Post } from 'src/posts/entities/post.entity';
+import { TownRange } from 'src/posts/entities/townRange.entity';
 import { User } from 'src/users/entities/user.entity';
 import { PriceOffer } from 'src/posts/priceOffer.entity';
-import { ComplaintReason } from 'src/complaintReasons/complaintReason.entity';
+import { ComplaintReason } from 'src/posts/entities/complaintReason.entity';
 import { PostComplaints } from 'src/posts/postComplaints.entity';
-import { ProcessState } from 'src/processStates/processState.entity';
+import { ProcessState } from 'src/posts/entities/processState.entity';
 import { PostsLikeRecord } from 'src/posts/postsLikeRecord.entity';
 import { PostsViewRecord } from 'src/posts/postsViewRecord.entity';
 import { Followings } from 'src/mypage/followings.entity';
