+++ src/mypage/dto/purchaseHistory.dto.ts
@@ -1,6 +1,6 @@
 import { Field, InputType } from '@nestjs/graphql';
 import { IsNotEmpty, IsNumber } from 'class-validator';
-import { Post } from '../../posts/post.entity';
+import { Post } from '../../posts/entities/post.entity';
 
 @InputType()
 export class PurchaseHistoryDto {
