+++ src/mypage/mypage.repository.ts
@@ -4,7 +4,7 @@ import { UpdateDealStateDto } from './../posts/dto/updateDealState.dto';
 import { EntityRepository, getRepository, Repository, EntityManager, getConnection } from 'typeorm';
 import { Followings } from './followings.entity';
 import { User } from 'src/users/entities/user.entity';
-import { Post } from 'src/posts/post.entity';
+import { Post } from 'src/posts/entities/post.entity';
 import { PurchaseHistory } from './purchaseHistory.entity';
 import { PurchaseHistoryDto } from './dto/purchaseHistory.dto';
