+++ src/posts/dto/addPostsLike.dto.ts
@@ -1,6 +1,6 @@
 import { Field, InputType } from '@nestjs/graphql';
 import { IsNotEmpty, IsNumber } from 'class-validator';
-import { Post } from '../post.entity';
+import { Post } from '../entities/post.entity';
 
 @InputType()
 export class PostsLikeDto {
