+++ src/posts/dto/createPost.dto.ts
@@ -1,9 +1,9 @@
 import { Field, InputType } from '@nestjs/graphql';
 import { IsBoolean, IsNotEmpty, IsNumber, IsString, Length } from 'class-validator';
 import * as GraphQLUpload from 'graphql-upload/GraphQLUpload.js';
-import { Category } from 'src/categories/category.entity';
-import { DealState } from 'src/dealStates/dealState.entity';
-import { TownRange } from 'src/townRanges/townRange.entity';
+import { Category } from 'src/posts/entities/category.entity';
+import { DealState } from 'src/posts/entities/dealState.entity';
+import { TownRange } from 'src/posts/entities/townRange.entity';
 import { FileUpload } from 'src/users/models/fileUpload.model';
 
 @InputType()
