+++ src/posts/dto/updateDealState.dto.ts
@@ -1,7 +1,7 @@
 import { Field, InputType } from '@nestjs/graphql';
 import { IsNotEmpty, IsNumber } from 'class-validator';
-import { Post } from '../post.entity';
-import { DealState } from 'src/dealStates/dealState.entity';
+import { Post } from '../entities/post.entity';
+import { DealState } from 'src/posts/entities/dealState.entity';
 
 @InputType()
 export class UpdateDealStateDto {
