+++ src/posts/entities/processState.entity.ts
@@ -1,8 +1,8 @@
-import { UserComplaints } from './../chats/userComplaints.entity';
-import { ChatComplaints } from './../chats/chatComplaints.entity';
+import { UserComplaints } from '../../chats/userComplaints.entity';
+import { ChatComplaints } from '../../chats/chatComplaints.entity';
 import { Field, ObjectType } from '@nestjs/graphql';
 import { BaseEntity, Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
-import { PostComplaints } from '../posts/postComplaints.entity';
+import { PostComplaints } from '../postComplaints.entity';
 
 @Entity()
 @ObjectType()
