+++ src/posts/postComplaints.entity.ts
@@ -1,8 +1,8 @@
 import { Field, ObjectType } from '@nestjs/graphql';
 import { BaseEntity, Column, CreateDateColumn, Entity, JoinColumn, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
-import { Post } from './post.entity';
-import { ProcessState } from 'src/processStates/processState.entity';
-import { ComplaintReason } from 'src/complaintReasons/complaintReason.entity';
+import { Post } from './entities/post.entity';
+import { ProcessState } from 'src/posts/entities/processState.entity';
+import { ComplaintReason } from 'src/posts/entities/complaintReason.entity';
 
 @Entity()
 @ObjectType()
