+++ src/posts/postsViewRecord.entity.ts
@@ -1,7 +1,7 @@
 import { Field, ObjectType } from '@nestjs/graphql';
 import { User } from 'src/users/entities/user.entity';
 import { BaseEntity, Entity, JoinColumn, PrimaryGeneratedColumn, ManyToOne } from 'typeorm';
-import { Post } from './post.entity';
+import { Post } from './entities/post.entity';
 
 @Entity()
 @ObjectType()
