+++ src/posts/priceOffer.entity.ts
@@ -1,6 +1,6 @@
 import { BaseEntity, Column, CreateDateColumn, Entity, PrimaryGeneratedColumn, ManyToOne, JoinColumn } from 'typeorm';
 import { Field, Int, ObjectType } from '@nestjs/graphql';
-import { Post } from './post.entity';
+import { Post } from './entities/post.entity';
 
 @Entity()
 @ObjectType()
