+++ src/posts/repositories/post.repository.ts
@@ -7,11 +7,11 @@ import { EntityManager, EntityRepository, getConnection, getRepository, Reposito
 import { CreatePostDto } from '../dto/createPost.dto';
 import { SearchPostDto } from '../dto/searchPost.dto';
 import { UpdatePostDto } from '../dto/updatePost.dto';
-import { Post } from '../post.entity';
+import { Post } from '../entities/post.entity';
 import { PriceOffer } from '../priceOffer.entity';
-import { ProcessState } from 'src/processStates/processState.entity';
+import { ProcessState } from 'src/posts/entities/processState.entity';
 import { PostComplaints } from '../postComplaints.entity';
-import { DealState } from 'src/dealStates/dealState.entity';
+import { DealState } from 'src/posts/entities/dealState.entity';
 import { User } from 'src/users/entities/user.entity';
 import { PostsLikeRecord } from '../postsLikeRecord.entity';
 import { PostsLikeDto } from '../dto/addPostsLike.dto';
